Fix uninstall

+++ base.mk
@@ -7,9 +7,7 @@ install-requirements:
 install-local-requirements:
 	python -m pip install --upgrade pip;
 	pip install -r requirements.txt -t ./;
-	if [ -f "uninstall.txt" ]; then
-	  pip uninstall -r uninstall.txt -y;
-	fi
+	test -f uninstall.txt && pip uninstall -r uninstall.txt -y;
 
 update-submodule:
 	git submodule update --remote --merge
